test: remove needless tests
For example, the following case focuses Ruby's basic feature (argument
check feature). It's not a test area of us.

    test("no values") do
      assert_raise(ArgumentError){ between("tags") }
    end

For example, the following case doesn't show expected behavior. We
expect String or Symbol as the first argument. We don't expect number as
the first argument. It's unexpected behavior (we can validate the first
argument or we just use the argument value because it's responsive of
user). We shouldn't add a test for the unexpected behavior. If we
validate the first argument then it's reasonable that we add a test
that the method raises for wrong value.

    test("number_include") do
      assert_equal({
                     :table => "posts",
                     :filter => "between(10, 2, \"include\", 29, \"include\")",
                   },
                   between(10, 2, "include", 29, "include"))
    end